C4-71 Filter Sets (#122)
* filter-set schema

* add types file

* add insert

* filter set added to mapping on deploy

* add queries to inserts

* filter_set added to test order

* add embedding

* C4-77 start refactoring into monolithic state class

* C4-77 VERY ROUGH initial refactor

* C4-71 fix bug in refactor

* C4-71 partway through larger refactor

* C4-71 another intermediary

* C4-71 complete initial refactor for filter sets

* C4-71 fix mock

* C4-71 more small cleanup

* C4-71 bootstrap filter_set with tests, execute action

* C4-71 sketch out compound_search

* C4-71 add some more functionality that will be needed

* C4-71 address some review points

* C4-71 another round of review changes

* C4-71 fix bug in generator search

* C4-71 prepare es_testapp

* C4-71 refactor a little, start implementing top down

* C4-71 configure how the frontend will interact

* C4-71 almost there

* C4-71 pull it together

* C4-71 add some variants to workbook_inserts for testing, clarify some comments

* C4-71 test some more

* C4-71 refactor compound search functions into class, add another test

* C4-71 quick documentation pass

* C4-71 small review changes

* C4-71 allow generator

* C4-71 add some more tests, small bug fixes

* C4-71 working pagination

* C4-71 optimize, more bug fixes

* C4-71 small updates, bit more testing

* C4-71 move split /compound_search from /search

* C4-71 tiny bug fix

* C4-71 address more comments

* C4-71 small review changes

* C4-191 fix add_default_schema_fields for genes

* C4-71 large batch of review changes

* C4-71 another batch of small review changes

* C4-71 fix docstring on route

* C4-71 more review updates

* fix missing import

* C4-71 last review changes, bump version to 2.1.0
