+++ pytest.ini
@@ -12,4 +12,5 @@ markers =
     slow: mark a test as slow (deselect with '-m "not slow"')
     storage: mark a test as about storage (deselect with '-m "not storage"')
     working: mark a test as working (deselect with '-m "not working"')
+    search: mark a test a search module test (deselect with '-m "not search"')
 norecursedirs = *env site-packages .cache .git .idea *.egg-info
