+++ src/encoded/tests/test_purge_item_type.py
@@ -1,6 +1,6 @@
 import pytest
 import time
-from .workbook_fixtures import app_settings, app, workbook
+from .workbook_fixtures import app, workbook
 from encoded.commands.purge_item_type import purge_item_type_from_storage
