+++ src/encoded/tests/workbook_fixtures.py
@@ -1,4 +1,3 @@
-import os
 import pkg_resources
 import pytest
 import webtest
@@ -7,7 +6,6 @@ from snovault import DBSESSION
 from snovault.elasticsearch import create_mapping
 from .. import main
 from ..loadxl import load_all
-from .conftest_settings import make_app_settings_dictionary
 
 
 # this file was previously used to setup the test fixtures for the BDD tests.
@@ -19,31 +17,12 @@ def external_tx():
     pass
 
 
-@pytest.fixture(scope='session')
-def app_settings(wsgi_server_host_port, elasticsearch_server, postgresql_server, aws_auth):
-    settings = make_app_settings_dictionary()
-    settings['create_tables'] = True
-    settings['persona.audiences'] = 'http://%s:%s' % wsgi_server_host_port
-    settings['elasticsearch.server'] = elasticsearch_server
-    settings['sqlalchemy.url'] = postgresql_server
-    settings['collection_datastore'] = 'elasticsearch'
-    settings['item_datastore'] = 'elasticsearch'
-    settings['indexer'] = True
-    settings['indexer.namespace'] = os.environ.get('TRAVIS_JOB_ID', '') # set namespace for tests
-
-    # use aws auth to access elasticsearch
-    if aws_auth:
-        settings['elasticsearch.aws_auth'] = aws_auth
-    return settings
-
-
 @pytest.yield_fixture(scope='session')
-def app(app_settings, **kwargs):
+def app(es_app_settings, **kwargs):
     """
     Pass all kwargs onto create_mapping
     """
-
-    app = main({}, **app_settings)
+    app = main({}, **es_app_settings)
     create_mapping.run(app, **kwargs)
 
     yield app
@@ -53,7 +32,7 @@ def app(app_settings, **kwargs):
     DBSession.bind.pool.dispose()
 
 
-@pytest.mark.fixture_cost(500)
+@pytest.mark.fixture_cost(500)  # XXX: this does nothing...? -will
 @pytest.yield_fixture(scope='session')
 def workbook(app):
     environ = {
